Refactor creation of `$nu.scope` in eval.rs
The function was ~400 lines long and hence very hard to work with.
