+++ crates/nu-engine/src/lib.rs
@@ -4,6 +4,7 @@ pub mod documentation;
 pub mod env;
 mod eval;
 mod glob_from;
+mod scope;
 
 pub use call_ext::CallExt;
 pub use column::get_columns;
